Extract OTP features to separate unit
