+++ NitrokeyManagerOTP.h
@@ -1,0 +1,6 @@
+#ifndef LIBNITROKEY_NITROKEYMANAGEROTP_H
+#define LIBNITROKEY_NITROKEYMANAGEROTP_H
+
+#include "NitrokeyManager.h"
+
+#endif // LIBNITROKEY_NITROKEYMANAGEROTP_H
